Add utils functions for direction and interpolation,
Added utility functions to handle direction and interpolation in the codebase. Updated imports and references accordingly.
