+++ lablib/operators/__init__.py
@@ -15,17 +15,26 @@ from .repositions import (
     Mirror2,
     CornerPin2D,
 )
+from .utils import (
+    get_direction,
+    get_interpolation,
+)
 
 __all__ = [
+    # Color
     "BaseOperator",
     "ColorOperator",
     "OCIOCDLTransform",
     "OCIOColorSpace",
     "OCIOFileTransform",
     "AYONOCIOLookProduct",
+    # Repositions
     "RepositionOperator",
     "Transform",
     "Crop",
     "Mirror2",
     "CornerPin2D",
+    # Utils
+    "get_direction",
+    "get_interpolation",
 ]
